Unhinged crap
why am I doing this why

+++ src/main/java/emi/lib/mtg/characteristic/ManaSymbol.java
@@ -6,7 +6,7 @@ import java.util.Collection;
  * Created by Emi on 5/6/2016.
  */
 public interface ManaSymbol {
-	int convertedCost();
+	double convertedCost();
 
 	boolean varies();
